requirements.txt
mcp-fastmcp>=0.1.0
typing>=3.7.4